Separated class and interface declarations into respective projects
