+++ test/TestGrainInterfaces/IGenericInterfaces.cs
@@ -230,4 +230,73 @@ namespace UnitTests.GrainInterfaces
 
 
 
+
+    namespace Generic.EdgeCases
+    {
+        public interface IBasicGrain : IGrainWithGuidKey
+        {
+            Task<string> Hello();
+            Task<string[]> ConcreteGenArgTypeNames();
+        }
+
+
+        public interface IGrainWithTwoGenArgs<T1, T2> : IBasicGrain
+        { }
+
+        public interface IGrainWithThreeGenArgs<T1, T2, T3> : IBasicGrain
+        { }
+
+        public interface IGrainReceivingRepeatedGenArgs<T1, T2> : IBasicGrain
+        { }
+
+        
+        public interface IPartiallySpecifyingInterface<T> : IGrainWithTwoGenArgs<T, int>
+        { }
+        
+
+        public interface IReceivingRepeatedGenArgsAmongstOthers<T1, T2, T3> : IBasicGrain
+        { }
+
+
+        public interface IReceivingRepeatedGenArgsFromOtherInterface<T1, T2, T3> : IBasicGrain
+        { }
+
+        public interface ISpecifyingGenArgsRepeatedlyToParentInterface<T> : IReceivingRepeatedGenArgsFromOtherInterface<T, T, T>
+        { }
+
+
+        public interface IReceivingRearrangedGenArgs<T1, T2> : IBasicGrain
+        { }
+
+
+
+        public interface IReceivingRearrangedGenArgsViaCast<T1, T2> : IBasicGrain
+        { }
+
+        public interface ISpecifyingRearrangedGenArgsToParentInterface<T1, T2> : IReceivingRearrangedGenArgsViaCast<T2, T1>
+        { }
+
+
+        public interface IArbitraryInterface<T1, T2> : IBasicGrain
+        { }
+
+        public interface IInterfaceUnrelatedToConcreteGenArgs<T> : IBasicGrain
+        { }
+
+
+        public interface IInterfaceTakingFurtherSpecializedGenArg<T> : IBasicGrain
+        { }
+
+
+        public interface IAnotherReceivingFurtherSpecializedGenArg<T> : IBasicGrain
+        { }
+
+        public interface IYetOneMoreReceivingFurtherSpecializedGenArg<T> : IBasicGrain
+        { }
+
+
+    }
+
+
+
 }
